Create post request for commenting on posts
