+++ html/feed.html
@@ -8,8 +8,9 @@
     <link rel="stylesheet" href="/dist/css/styles.css" />
     <script src="https://cdnjs.cloudflare.com/ajax/libs/dompurify/3.0.6/purify.min.js" integrity="sha512-H+rglffZ6f5gF7UJgvH4Naa+fGCgjrHKMgoFOGmcPTRwR6oILo5R+gtzNrpDp7iMV3udbymBVjkeZGNz1Em4rQ==" crossorigin="anonymous" referrerpolicy="no-referrer"></script>
     <script src="../src/scss/bootstrap.bundle.min.js" defer></script>
-    <script src="../js/apiCall.js" defer></script>
-    <script src="../js/postReactions.js" defer></script>
+    <script src="../js/apiCall.js" defer type="module"></script>
+    <script src="../js/postReactions.js" defer type="module"></script>
+    <script src="../js/postComment.js" defer type="module"></script>
 
 
   </head>
